Add draft Snakemake implementation
Rather than use the plot and rename applications, it is more idiomatic
in Snakemake to use R and Python (respectively) scripts that directly
reference the Snakemake objects holding inputs, outputs and other
values. This simplifies the R/Python code (no need to write a CLI), but
makes the code Snakemake-specific.